* DefaultSteps class splitted into ElementVerificationSteps, ElementInteractionSteps, InputFieldSteps, ListSteps and WebPageSteps * DefaultStepsTest class splitted (analogically) * ReadMe updated
