+++ README.md
@@ -111,6 +111,14 @@ public HeaderBlock header;
 ```
 При загрузке страницы будут учитываться элементы, описанные в блоке
 
+Хранилище переменных
+============================
+В akitaScenario используется хранилище переменных. Для сохранения/изъятия переменных используются методы setVar/getVar
+Каждая страница, с которой предполагается взаимодействие, должна быть описана в соответствующем классе,
+наследующем AkitaPage. Для каждого элемента следует задать имя на русском, через аннотацию @Name, чтобы искать
+можно было именно по русскому описанию, а не по селектору. Селекторы следует хранить только в классе страницы,
+не в степах, в степах - взаимодействие по русскому названию элемента.
+
 Screenshots
 ============================
 Реализован кастомный StepFormatter. При подключении его к проеку с тестами, становится достуна опция снятия скриншотов после желаемого или каждого шага. ```@CucumberOptions(format = {"pretty", "ru.alfabank.tests.core.formatters.StepFormatter"})```
